#74: Refactor - moved trajectory-services
