+++ src/app/debug-window/debug-window.component.spec.ts
@@ -5,7 +5,7 @@ import { LocalNotifications } from '@ionic-native/local-notifications/ngx'
 import { IonicModule } from '@ionic/angular'
 import { SqliteService } from '../shared-services/db/sqlite.service'
 import { LocationService } from '../shared-services/location.service'
-import { TrajectoryService } from '../shared-services/trajectory.service'
+import { TrajectoryService } from '../shared-services/trajectory/trajectory.service'
 
 import { DebugWindowComponent } from './debug-window.component'
