+++ src/app/select-trajectory/select-trajectory.page.spec.ts
@@ -7,8 +7,8 @@ import { SocialSharing } from '@ionic-native/social-sharing/ngx'
 import { IonicModule, IonRouterOutlet } from '@ionic/angular'
 import { SqliteService } from '../shared-services/db/sqlite.service'
 import { LocationService } from '../shared-services/location.service'
-import { TrajectoryImportExportService } from '../shared-services/trajectory-import-export.service'
-import { TrajectoryService } from '../shared-services/trajectory.service'
+import { TrajectoryImportExportService } from '../shared-services/trajectory/trajectory-import-export.service'
+import { TrajectoryService } from '../shared-services/trajectory/trajectory.service'
 import { SelectTrajectoryPage } from './select-trajectory.page'
 
 describe('SelectTrajectoryPage', () => {
