+++ src/app/select-trajectory/trajectory-card/trajectory-card-popover/trajectory-card-popover.page.spec.ts
@@ -7,8 +7,8 @@ import { SocialSharing } from '@ionic-native/social-sharing/ngx'
 import { IonicModule } from '@ionic/angular'
 import { SqliteService } from 'src/app/shared-services/db/sqlite.service'
 import { LocationService } from 'src/app/shared-services/location.service'
-import { TrajectoryImportExportService } from 'src/app/shared-services/trajectory-import-export.service'
-import { TrajectoryService } from 'src/app/shared-services/trajectory.service'
+import { TrajectoryImportExportService } from 'src/app/shared-services/trajectory/trajectory-import-export.service'
+import { TrajectoryService } from 'src/app/shared-services/trajectory/trajectory.service'
 import { TrajectoryCardPopoverPage } from './trajectory-card-popover.page'
 
 describe('TrajectoryCardPopoverPage', () => {
