+++ src/app/select-trajectory/trajectory-card/trajectory-card.component.spec.ts
@@ -6,7 +6,7 @@ import { LocalNotifications } from '@ionic-native/local-notifications/ngx'
 import { IonicModule } from '@ionic/angular'
 import { SqliteService } from 'src/app/shared-services/db/sqlite.service'
 import { LocationService } from 'src/app/shared-services/location.service'
-import { TrajectoryService } from 'src/app/shared-services/trajectory.service'
+import { TrajectoryService } from 'src/app/shared-services/trajectory/trajectory.service'
 import { TrajectoryCardComponent } from './trajectory-card.component'
 
 describe('TrajectoryCardComponent', () => {
