+++ src/app/shared-services/trajectory/trajectory-import-export.service.spec.fixtures.ts
@@ -3,7 +3,7 @@ import {
   TrajectoryData,
   TrajectoryMeta,
   TrajectoryType,
-} from '../model/trajectory'
+} from '../../model/trajectory'
 
 export const testTrajectoryName = 'test-name'
