+++ src/app/shared-services/trajectory/trajectory.service.spec.ts
@@ -2,8 +2,8 @@ import { HttpClientTestingModule } from '@angular/common/http/testing'
 import { TestBed } from '@angular/core/testing'
 import { BackgroundGeolocation } from '@ionic-native/background-geolocation/ngx'
 import { LocalNotifications } from '@ionic-native/local-notifications/ngx'
-import { SqliteService } from './db/sqlite.service'
-import { LocationService } from './location.service'
+import { SqliteService } from './../db/sqlite.service'
+import { LocationService } from './../location.service'
 import { TrajectoryService } from './trajectory.service'
 
 describe('TrajectoryService', () => {
