+++ src/app/tracking/tracking.page.ts
@@ -4,7 +4,7 @@ import { Platform } from '@ionic/angular'
 import { Subscription } from 'rxjs'
 import { TrajectoryType } from '../model/trajectory'
 import { LocationService } from '../shared-services/location.service'
-import { TrajectoryService } from '../shared-services/trajectory.service'
+import { TrajectoryService } from '../shared-services/trajectory/trajectory.service'
 
 @Component({
   selector: 'app-tracking',
