+++ src/app/trajectory/map/map.page.spec.ts
@@ -7,7 +7,7 @@ import { LocalNotifications } from '@ionic-native/local-notifications/ngx'
 import { IonicModule } from '@ionic/angular'
 import { SqliteService } from '../../shared-services/db/sqlite.service'
 import { LocationService } from '../../shared-services/location.service'
-import { TrajectoryService } from '../../shared-services/trajectory.service'
+import { TrajectoryService } from '../../shared-services/trajectory/trajectory.service'
 import { MapPage } from './map.page'
 
 describe('MapPage', () => {
